Separate theme into separate file
